Initialize a book template with named chapters and base content

Initialize a book template with named chapters and base content

initialize_template(
  path,
  chapters = c("Preface", "Introduction"),
  references = "References"
)

Arguments

path
where to create the template

chapters
titles of Rmd files to include in the template

references
title of the references Rmd or NULL

Value

used for side effect: file creation in path

Examples

# Create temporary directory for reproducible example
dir_tmp <- tempfile(pattern = "proj-")
dir.create(dir_tmp)
# browseURL(dir_tmp)

initialize_template(
  path = dir_tmp,
  chapters = c(
    "Introduction", "Material and Methods",
    "Results", "Discussion"
  ),
  references = NULL
)